Update README.md & workflow images - add new version argument.
_______________________________
DevNote
a thought just occurred to me, of a likely blunder in nice block size.
there should, at the very minimum, be an addressable area of size equal to the lowest "best" multiple; and the bigger the addressable area the more variation allowed. Therefore, instead of aiming for a highest multiple to maximize performance, I should take this into consideration and further limit the maximum possible value, or define some criteria to better balance the decision.